Update how table prefix columns are scanned into structs
The previous way was too ambigious for my liking, having to define the
tag as "users.*". This new way, is more inline with how sub-struct
mapping is handled anyway. Now, when working with joined columns the
user has to specify "users.*:*". This makes it more consistent with what
is already done for sub-struct fields, for example "user_id:id".

This also means that column names that a prefixed can be mapped to
regular struct fields, for example,

    type Post struct {
        User *User `db:"users.id:id"`
    }

This provides more clarity around the columns that can be mapped.